+++ projects/ontimize-web-ngx/src/lib/components/contextmenu/context-menu-separator/o-context-menu-separator.component.ts
@@ -1,4 +1,4 @@
-import { Component, forwardRef, OnInit } from '@angular/core';
+import { Component, forwardRef } from '@angular/core';
 
 import { OContextMenuItemComponent } from '../context-menu-item/o-context-menu-item.component';
 import { DEFAULT_INPUTS_O_CONTEXT_MENU_ITEMS, OComponentMenuItems } from '../o-content-menu.class';
@@ -13,12 +13,8 @@ export const DEFAULT_CONTEXT_MENU_ITEM_INPUTS = [
   inputs: DEFAULT_CONTEXT_MENU_ITEM_INPUTS,
   providers: [{ provide: OComponentMenuItems, useExisting: forwardRef(() => OContextMenuSeparatorComponent) }]
 })
-export class OContextMenuSeparatorComponent extends OContextMenuItemComponent implements OnInit {
+export class OContextMenuSeparatorComponent extends OContextMenuItemComponent  {
 
   public type = OComponentMenuItems.TYPE_SEPARATOR_MENU;
-
-  public ngOnInit(): void {
-    super.ngOnInit();
-  }
 
 }