+++ projects/ontimize-web-ngx/src/lib/components/table/column/cell-editor/o-base-table-cell-editor.class.ts
@@ -85,6 +85,9 @@ export class OBaseTableCellEditor implements OnInit {
     this.initialize();
   }
 
+  /**
+   * Initializes the cell editor
+   */
   public initialize(): void {
     this.createFormControl();
     this.registerEditor();
@@ -96,15 +99,29 @@ export class OBaseTableCellEditor implements OnInit {
     if (!oColumn || !oColumn.editing) {
       return;
     }
-    if (event.keyCode === 27) {
-      // escape
+
+    if (this.checkKey(event, "Escape", 27)) {
       this.onEscClicked();
-    } else if (event.keyCode === 13 || event.keyCode === 9) {
+      return;
+    }
+
+    if (!this.table.editingCell.contains(event.target)) {
+      return;
+    }
+
+    if (this.checkKey(event, "Enter", 13) || this.checkKey(event, "Tab", 9)) {
       // intro or tab
       this.commitEdition();
     }
   }
 
+  protected checkKey(event: KeyboardEvent, key: string, keyCode: number): boolean {
+    return (event.key && event.key === key) || (event.keyCode && event.keyCode === keyCode);
+  }
+
+  /**
+   * Creates form control
+   */
   createFormControl() {
     if (!this.formControl) {
       const validators: ValidatorFn[] = this.resolveValidators();
@@ -117,6 +134,9 @@ export class OBaseTableCellEditor implements OnInit {
     }
   }
 
+  /**
+   * Registers editor
+   */
   registerEditor() {
     if (this.registerInColumn && !Util.isDefined(this.tableColumn.editor)) {
       this.tableColumn.registerEditor(this);
@@ -126,10 +146,18 @@ export class OBaseTableCellEditor implements OnInit {
     }
   }
 
+  /**
+   * Gets the value of the cell data
+   * @returns cell data
+   */
   getCellData(): any {
     return this._rowData[this.tableColumnAttr];
   }
 
+  /**
+   * Start edition with given the data
+   * @param data
+   */
   startEdition(data: any) {
     this.formGroup.reset();
     this.rowData = data;
@@ -138,7 +166,13 @@ export class OBaseTableCellEditor implements OnInit {
     }
   }
 
-  endEdition(saveChanges) {
+
+
+  /**
+   * Ends edition with the ability to skip or save changes
+   * @param saveChanges
+   */
+  endEdition(saveChanges: boolean) {
     const oColumn: OColumn = this.table.getOColumn(this.tableColumnAttr);
     if (oColumn) {
       const self = this;
@@ -159,6 +193,10 @@ export class OBaseTableCellEditor implements OnInit {
     }
   }
 
+
+  /**
+   * Commits edition
+   */
   commitEdition() {
     if (!this.formControl.invalid) {
       this.oldValue = this._rowData[this.tableColumnAttr];
@@ -211,6 +249,11 @@ export class OBaseTableCellEditor implements OnInit {
     }
   }
 
+
+  /**
+   * Resolves validators
+   * @returns validators
+   */
   resolveValidators(): ValidatorFn[] {
     const validators: ValidatorFn[] = [];
     if (this.orequired) {
@@ -219,6 +262,12 @@ export class OBaseTableCellEditor implements OnInit {
     return validators;
   }
 
+
+  /**
+   * Determines whether error has
+   * @param error
+   * @returns true if error
+   */
   hasError(error: string): boolean {
     return this.formControl && this.formControl.touched && this.hasErrorExclusive(error);
   }