+++ projects/ontimize-web-ngx/src/lib/interfaces/o-table-column.interface.ts
@@ -13,6 +13,7 @@ export interface OTableColumn {
   multiline: boolean;
   orderable: boolean;
   searchable: boolean;
+  groupable: boolean;
   resizable: boolean;
   titleAlign: string;
   contentAlign: string;