+++ projects/ontimize-web-ngx/src/lib/layouts/app-layout/o-app-layout.component.ts
@@ -1,6 +1,10 @@
-import { Component, EventEmitter, ViewChild, ViewEncapsulation } from '@angular/core';
+import { Component, ContentChild, EventEmitter, ViewChild, ViewEncapsulation } from '@angular/core';
+import { OAppHeaderComponent } from '../../components/app-header/o-app-header.component';
+import { ThemePalette } from '@angular/material';
+
 
 import { OAppSidenavComponent } from '../../components/app-sidenav/o-app-sidenav.component';
+import { OUserInfoConfigurationDirective } from '../../components/user-info/user-info-configuration/o-user-info-configuration.directive';
 import { InputConverter } from '../../decorators/input-converter';
 import { Codes, OAppLayoutMode, OSidenavMode } from '../../util/codes';
 import { Util } from '../../util/util';
@@ -14,7 +18,8 @@ export const DEFAULT_INPUTS_O_APP_LAYOUT = [
   'showLanguageSelector: show-language-selector',
   'useFlagIcons: use-flag-icons',
   'openedSidenavImg: opened-sidenav-image',
-  'closedSidenavImg: closed-sidenav-image'
+  'closedSidenavImg: closed-sidenav-image',
+  'headerColor: header-color'
 ];
 
 export const DEFAULT_OUTPUTS_O_APP_LAYOUT: any[] = [
@@ -46,8 +51,16 @@ export class OAppLayoutComponent {
   @InputConverter()
   protected _showHeader: boolean;
 
+  public headerColor: ThemePalette;
+
   @ViewChild('appSidenav', { static: false })
   public appSidenav: OAppSidenavComponent;
+
+  @ViewChild('appHeader', { static: false })
+  public appHeader: OAppHeaderComponent;
+
+  @ContentChild(OUserInfoConfigurationDirective, { static: false })
+  public userInfoConfiguration: OUserInfoConfigurationDirective;
 
   protected _mode: OAppLayoutMode;
   protected _sidenavMode: OSidenavMode;
@@ -104,4 +117,11 @@ export class OAppLayoutComponent {
   afterToggle(opened: boolean) {
     opened ? this.afterOpenSidenav.emit() : this.afterCloseSidenav.emit();
   }
+
+  ngAfterViewInit(): void {
+    if (this.appHeader && this.appHeader.userInfo) {
+      this.appHeader.userInfo.registerUserInfoConfiguration(this.userInfoConfiguration);
+    }
+  }
+
 }